+++ stdlib/lib.core.d.ts
@@ -1987,6 +1987,7 @@ interface EventsModule {
     listenerCount(emitter: EventEmitter, eventName: string, listener?: (this: EventEmitter, ...args: any[]) => void, ...ignored: any[]): number;
     getEventListeners(emitter: EventEmitter, eventName: string, ...ignored: any[]): any[];
     once(emitter: EventEmitter, eventName: string, options?: EventEmitterOnceOptions, ...ignored: any[]): Promise<any[]>;
+    on(emitter: EventEmitter, eventName: string, ...ignored: any[]): any;
     setMaxListeners(n: number, emitter: EventEmitter, ...emitters: EventEmitter[]): void;
     getMaxListeners(emitter: EventEmitter, ...ignored: any[]): number;
 }
@@ -1996,6 +1997,7 @@ declare module "events" {
     export function listenerCount(emitter: EventEmitter, eventName: string, listener?: (this: EventEmitter, ...args: any[]) => void, ...ignored: any[]): number;
     export function getEventListeners(emitter: EventEmitter, eventName: string, ...ignored: any[]): any[];
     export function once(emitter: EventEmitter, eventName: string, options?: EventEmitterOnceOptions, ...ignored: any[]): Promise<any[]>;
+    export function on(emitter: EventEmitter, eventName: string, ...ignored: any[]): any;
     export function setMaxListeners(n: number, emitter: EventEmitter, ...emitters: EventEmitter[]): void;
     export function getMaxListeners(emitter: EventEmitter, ...ignored: any[]): number;
     const defaultEvents: EventsModule;
@@ -2007,6 +2009,7 @@ declare module "node:events" {
     export function listenerCount(emitter: EventEmitter, eventName: string, listener?: (this: EventEmitter, ...args: any[]) => void, ...ignored: any[]): number;
     export function getEventListeners(emitter: EventEmitter, eventName: string, ...ignored: any[]): any[];
     export function once(emitter: EventEmitter, eventName: string, options?: EventEmitterOnceOptions, ...ignored: any[]): Promise<any[]>;
+    export function on(emitter: EventEmitter, eventName: string, ...ignored: any[]): any;
     export function setMaxListeners(n: number, emitter: EventEmitter, ...emitters: EventEmitter[]): void;
     export function getMaxListeners(emitter: EventEmitter, ...ignored: any[]): number;
     const defaultEvents: EventsModule;